extract paste handler into separate hook
