+++ src/components/MessageInput/hooks/useMessageInputState.ts
@@ -1,5 +1,5 @@
 import React, { Reducer, useCallback, useReducer } from 'react';
-import { dataTransferItemsHaveFiles, dataTransferItemsToFiles, FileLike } from 'react-file-utils';
+import type { FileLike } from 'react-file-utils';
 import type { Attachment, UserResponse } from 'stream-chat';
 
 import { StreamMessage, useChannelStateContext } from '../../../context/ChannelStateContext';
@@ -10,6 +10,7 @@ import { useAttachments } from './useAttachments';
 import { useMessageInputText } from './useMessageInputText';
 import { useEmojiPicker } from './useEmojiPicker';
 import { useSubmitHandler } from './useSubmitHandler';
+import { usePasteHandler } from './usePasteHandler';
 
 import type { EmojiData, NimbleEmojiIndex } from 'emoji-mart';
 
@@ -374,45 +375,7 @@ export const useMessageInputState = <
     uploadNewFiles,
   } = useAttachments(props, state, dispatch);
 
-  const onPaste = useCallback(
-    (clipboardEvent: React.ClipboardEvent<HTMLTextAreaElement>) => {
-      (async (event) => {
-        // TODO: Move this handler to package with ImageDropzone
-        const { items } = event.clipboardData;
-        if (!dataTransferItemsHaveFiles(Array.from(items))) return;
-
-        event.preventDefault();
-        // Get a promise for the plain text in case no files are
-        // found. This needs to be done here because chrome cleans
-        // up the DataTransferItems after resolving of a promise.
-        let plainTextPromise: Promise<string> | undefined = undefined;
-        for (let i = 0; i < items.length; i++) {
-          const item = items[i];
-          if (item.kind === 'string' && item.type === 'text/plain') {
-            plainTextPromise = new Promise((resolve) => {
-              item.getAsString((string) => {
-                resolve(string);
-              });
-            });
-            break;
-          }
-        }
-
-        const fileLikes = await dataTransferItemsToFiles(Array.from(items));
-        if (fileLikes.length) {
-          uploadNewFiles(fileLikes);
-          return;
-        }
-
-        // fallback to regular text paste
-        if (plainTextPromise) {
-          const pastedText = await plainTextPromise;
-          insertText(pastedText);
-        }
-      })(clipboardEvent);
-    },
-    [insertText, uploadNewFiles],
-  );
+  const { onPaste } = usePasteHandler(uploadNewFiles, insertText);
 
   const isUploadEnabled = channel?.getConfig?.()?.uploads !== false;
